// honey-paradise-client/src/shared/types/server/getMySettingsResponseDefaultTheme.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Honey Paradise
 * Honey Paradise API
 * OpenAPI spec version: 0.0.2
 */

export type GetMySettingsResponseDefaultTheme = typeof GetMySettingsResponseDefaultTheme[keyof typeof GetMySettingsResponseDefaultTheme];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetMySettingsResponseDefaultTheme = {
  DARK: 'DARK',
  LIGHT: 'LIGHT',
} as const;