// honey-paradise-client/src/shared/types/server/getMySettingsResponse.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Honey Paradise
 * Honey Paradise API
 * OpenAPI spec version: 0.0.2
 */
import type { GetMySettingsResponseDefaultLanguage } from './getMySettingsResponseDefaultLanguage';
import type { GetMySettingsResponseDefaultTheme } from './getMySettingsResponseDefaultTheme';
import type { GetMySettingsResponseDefaultCurrency } from './getMySettingsResponseDefaultCurrency';

export interface GetMySettingsResponse {
  id: string;
  defaultLanguage: GetMySettingsResponseDefaultLanguage;
  defaultTheme: GetMySettingsResponseDefaultTheme;
  defaultCurrency: GetMySettingsResponseDefaultCurrency;
  useFullLogout: boolean;
  useTgTfaLogin: boolean;
  createdAt: string;
  updatedAt: string;
}